refactor: Extract section grouping dataset to its own file
