+++ app/src/test/java/com/infomaniak/swisstransfer/TransferSectionGroupingTest.kt
@@ -18,6 +18,13 @@
 package com.infomaniak.swisstransfer
 
 import com.infomaniak.multiplatform_swisstransfer.common.interfaces.ui.TransferUi
+import com.infomaniak.swisstransfer.dataset.TransferSectionGroupingData.december
+import com.infomaniak.swisstransfer.dataset.TransferSectionGroupingData.future
+import com.infomaniak.swisstransfer.dataset.TransferSectionGroupingData.lastWeek
+import com.infomaniak.swisstransfer.dataset.TransferSectionGroupingData.thisMonth
+import com.infomaniak.swisstransfer.dataset.TransferSectionGroupingData.thisWeek
+import com.infomaniak.swisstransfer.dataset.TransferSectionGroupingData.today
+import com.infomaniak.swisstransfer.dataset.TransferSectionGroupingData.yesterday
 import com.infomaniak.swisstransfer.ui.screen.main.transfers.TransfersGroupingManager.SpecificSection
 import com.infomaniak.swisstransfer.ui.screen.main.transfers.TransfersGroupingManager.TransferSection
 import com.infomaniak.swisstransfer.ui.screen.main.transfers.TransfersGroupingManager.getMonthNameFromEpoch
@@ -88,88 +95,6 @@ class TransferSectionGroupingTest {
     }
 
     private companion object {
-        val december = listOf(
-            1734424201L, // 2024-12-17 TUESDAY
-            1734510601L, // 2024-12-18 WEDNESDAY
-            1734597001L, // 2024-12-19 THURSDAY
-            1734683401L, // 2024-12-20 FRIDAY
-            1734769801L, // 2024-12-21 SATURDAY
-            1734856201L, // 2024-12-22 SUNDAY
-            1734942601L, // 2024-12-23 MONDAY
-            1735029001L, // 2024-12-24 TUESDAY
-            1735115401L, // 2024-12-25 WEDNESDAY
-            1735201801L, // 2024-12-26 THURSDAY
-            1735288201L, // 2024-12-27 FRIDAY
-            1735374601L, // 2024-12-28 SATURDAY
-            1735461001L, // 2024-12-29 SUNDAY
-            1735547401L, // 2024-12-30 MONDAY
-            1735633801L, // 2024-12-31 TUESDAY
-        )
-
-        val thisMonth = listOf(
-            1735720201L, // 2025-01-01 WEDNESDAY
-            1735806601L, // 2025-01-02 THURSDAY
-            1735893001L, // 2025-01-03 FRIDAY
-            1735979401L, // 2025-01-04 SATURDAY
-            1736065801L, // 2025-01-05 SUNDAY
-        )
-
-        val lastWeek = listOf(
-            1736152201L, // 2025-01-06 MONDAY
-            1736238601L, // 2025-01-07 TUESDAY
-            1736325001L, // 2025-01-08 WEDNESDAY
-            1736411401L, // 2025-01-09 THURSDAY
-            1736497801L, // 2025-01-10 FRIDAY
-            1736584201L, // 2025-01-11 SATURDAY
-            1736670601L, // 2025-01-12 SUNDAY
-        )
-
-        val thisWeek = listOf(
-            1736757001L, // 2025-01-13 MONDAY
-            1736843401L, // 2025-01-14 TUESDAY
-        )
-
-        val yesterday = listOf(
-            1736929801L, // 2025-01-15 WEDNESDAY
-        )
-
-        val today = listOf(
-            1737016201L, // 2025-01-16 THURSDAY
-        )
-
-        val future = listOf(
-            1737102601L, // 2025-01-17 FRIDAY
-            1737189001L, // 2025-01-18 SATURDAY
-            1737275401L, // 2025-01-19 SUNDAY
-            1737361801L, // 2025-01-20 MONDAY
-            1737448201L, // 2025-01-21 TUESDAY
-            1737534601L, // 2025-01-22 WEDNESDAY
-            1737621001L, // 2025-01-23 THURSDAY
-            1737707401L, // 2025-01-24 FRIDAY
-            1737793801L, // 2025-01-25 SATURDAY
-            1737880201L, // 2025-01-26 SUNDAY
-            1737966601L, // 2025-01-27 MONDAY
-            1738053001L, // 2025-01-28 TUESDAY
-            1738139401L, // 2025-01-29 WEDNESDAY
-            1738225801L, // 2025-01-30 THURSDAY
-            1738312201L, // 2025-01-31 FRIDAY
-            1738398601L, // 2025-02-01 SATURDAY
-            1738485001L, // 2025-02-02 SUNDAY
-            1738571401L, // 2025-02-03 MONDAY
-            1738657801L, // 2025-02-04 TUESDAY
-            1738744201L, // 2025-02-05 WEDNESDAY
-            1738830601L, // 2025-02-06 THURSDAY
-            1738917001L, // 2025-02-07 FRIDAY
-            1739003401L, // 2025-02-08 SATURDAY
-            1739089801L, // 2025-02-09 SUNDAY
-            1739176201L, // 2025-02-10 MONDAY
-            1739262601L, // 2025-02-11 TUESDAY
-            1739349001L, // 2025-02-12 WEDNESDAY
-            1739435401L, // 2025-02-13 THURSDAY
-            1739521801L, // 2025-02-14 FRIDAY
-            1739608201L, // 2025-02-15 SATURDAY
-        )
-
         val groupedTransferUi = listOf(december, lastWeek, thisMonth, thisWeek, yesterday, today, future)
             .flatten()
             .sortedDescending()
